Add SV Raid Pointer 3.0.0

+++ SysBot.Pokemon/SV/Vision/PokeDataOffsetsSV.cs
@@ -30,8 +30,8 @@ namespace SysBot.Pokemon
         public IReadOnlyList<long> RaidBlockPointerB { get; } = new long[] { 0x47350D8, 0x1C0, 0x88, 0x1958 };
         public IReadOnlyList<long> BlockKeyPointer { get; } = new long[] { 0x47350D8, 0xD8, 0x0, 0x0, 0x30, 0x0 };
 
-        public uint TeraLobbyIsConnected { get; } = 0x043DF430;
-        public uint LoadedIntoDesiredState { get; } = 0x047D2020;
+        public ulong TeraLobbyIsConnected { get; } = 0x043DF430;
+        public ulong LoadedIntoDesiredState { get; } = 0x047D2020;
         public ulong MyNIDPointer { get; } = 0x0475E938; // MyNid
     }
 }
